Add files via upload
Remix this Glitch, or alternatively, recreate the project with local files on your own computer.
Notice there is a secretsRouter.js module in the /routes folder. In the server.js file, require the secretsRouter module, then use it to handle requests to the '/secrets' path. 
In secretsRouter.js, create and export an Express router that handles GET requests to '/'. This should send back a message of some kind, formatted as HTML. It can be any secret message you like. 
In server.js, use what you have learned about routing and Basic Authentication to secure this route until a username of jbond and the password AstonMartin007 have been entered. The username check should be case insensitive. The password check should be case sensitive. The secrets route should return a status code of 401 and string of "You are not authorized to view this resource" if a call is made without authorization or the wrong username and password. 
Testing
Open your Glitch in a incognito/private browser window. Use the Show -> In a New Window option, then add /secrets to the end of the path in the browser address bar. You should be challenged to enter your credentials.
Test with wrong credentials, including using the right password but with incorrect casing to make sure that your case sensitivity check for the password is working, as well as entering blank credentials or trying to cancel from the authentication challenge dialog.
Ensure that when you do enter the correct credentials, you are able to see your secret message.
To complete this Code Challenge, respond with a post containing either the link to your remixed Glitch, or if you've chosen to use local files on your computer instead, either attach your files or copy/paste the code from the files into your post.